+++ ase-aat/src/main/java/de/tum/aat/dao/StudentDAO.java
@@ -6,7 +6,9 @@ import java.util.List;
 
 import com.googlecode.objectify.Key;
 
+import de.tum.aat.domain.ExerciseGroup;
 import de.tum.aat.domain.Student;
+import de.tum.aat.exceptions.GenericException;
 import de.tum.aat.exceptions.NotFoundException;
 
 public class StudentDAO {
@@ -39,6 +41,11 @@ public class StudentDAO {
 			throw new NotFoundException(Student.class);
 		}
 
+		if (s.getExerciseGroup() != null) {
+			throw new GenericException("Student is enrolled in the " + s.getExerciseGroup().getName()
+					+ ". You can't delete registered student.");
+		}
+
 		ofy().delete().key(key).now();
 	}
 
@@ -51,6 +58,30 @@ public class StudentDAO {
 
 		oldStudent = s;
 		return saveStudent(oldStudent);
+	}
+
+	public Student registerExerciseGroupForStudent(long groupId, long studentId) {
+		Key<ExerciseGroup> groupKey = Key.create(ExerciseGroup.class, groupId);
+		Key<Student> studentKey = Key.create(Student.class, studentId);
+
+		ExerciseGroup g = ofy().load().key(groupKey).now();
+		Student s = ofy().load().key(studentKey).now();
+
+		if (s == null) {
+			throw new NotFoundException(Student.class);
+		}
+
+		if (g == null) {
+			throw new NotFoundException(ExerciseGroup.class);
+		}
+
+		s.setExerciseGroup(groupKey);
+
+		g.getStudents().add(s);
+
+		ofy().save().entity(g).now();
+
+		return saveStudent(s);
 	}
 
 }
